Add files via upload
This dashboard provides insight on drive data obtained from BackBlaze. I wanted to look at the Hitachi brand and what the differences would be in different models. After comparing the models for failure, I decided to look a bit further and compare the model with the most failures to the model with the least. The drive with the most failures actually had a larger sample size then the model with the least which could potentially be why there was a significant difference in the count of failures.